Try to support launching from Dock
This isn't the ideal way to support this (the right way would be to package
everything up with electron-builder and distribute a dmg) but it will
minimise surprise when people install from github, click "Keep in dock"
and then relaunch, preventing it from just opening the Electron demo
app.

It does mean the PATH and cwd are different from launching from the
terminal, so some workarounds are needed for this.